Specify that our counter is unsigned.
(Silencing a warning)

+++ KeyboardioScanner.cpp
@@ -130,7 +130,7 @@ void KeyboardioScanner::sendLEDData() {
 void KeyboardioScanner::sendLEDBank(byte bank) {
     Wire.beginTransmission(addr);
     Wire.write(TWI_CMD_LED_BASE+bank);
-    for (int i=0; i<LED_BYTES_PER_BANK; i++) {
+    for (uint8_t i=0; i<LED_BYTES_PER_BANK; i++) {
         Wire.write(ledData.bytes[bank][i]);
     }
     Wire.endTransmission();
